+++ src/balance.d.ts
@@ -1,0 +1,4 @@
+export function setBalancingParameters(
+  delta: number,
+  ratio: number,
+): undefined;