+++ src/index.d.ts
@@ -1,4 +1,5 @@
 export {default as at} from './at';
+export {setBalancingParameters} from './balance';
 export {default as create} from './create';
 export {default as difference} from './difference';
 export {default as equals} from './equals';